## ConsoleUtil
The Methods `lock` and `lockWithKeyCodes` where added.
This locks the Console should only be used at start if one wants to add the console as hidden element when embeded.


## AbstractConsole

Added a `ConsoleLock` element on it which is the lock used for indicating wether it is a locked console or not.
 

## Dlogger/DConsole

In their `toggleDisplay` they make sure to return early if they `lock` is locked.


## ConsoleLock

Added as the first element of the `security` package just a custom lock managed by the AbstractConsole and it's subclasses.


## KeyboardManager

Exposes now a shortcut method on the `validateKeyboardSequence` method from the `KeyboardSequences class.`


## KeyboardSequences

* Fixed and bug in the in the private `inKeyboardSequence` method which tried was accessing
the keyCodes on the object wrongly.

* A private `executeCallback` method was introduced  to be called from  the`onKeyUp` method, callbacks are now executed outside the iterator loop since external objects may remove a element from the KeyboardSequences when the `callback` is triggered.

* The onKeyUp does not track SHIFT/ALT/CTRL as keyCodes, charCode is default but tracks
keyCode when the flags SHIFT/ALT/CTRL are set.

## KeyboardBindings

* Shortcuts to the CharCodes for the lowercase variant has been added.